# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Timefly API
  description: MongoDB ExpressJS NodeJS REST API for a Project Management App.<br />Github repository available [here](https://github.com/aneguet/MEVN_ProjectManagementApp).<br />Some endpoints require `auth-token` in order to be executed.
  version: 1.0.0
servers:
  - url: https://timefly-mevn.herokuapp.com/api/
    description: Heroku remote server
  - url: http://localhost:4000/api/
    description: Localhost
components:
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: auth-token
  schemas:
...
    Project:
      type: object
      required:
        - 'name'
        - 'description'
        - 'stakeholder'
        - 'technologies'
        - 'time_schedule'
      properties:
        name:
          type: string
        description:
          type: string
        stakeholder:
          type: string
          example: 'Mercedes Benz'
        leader:
          type: string
          description: 'Must be an id composed by numbers and letters'
        members:
          type: array
          description: 'Array of members ids'
          items:
            type: object
            properties:
              member_id:
                type: string
              project_roles:
                type: array
                items:
                  type: string
                  default: 'Timefly member'
                  example: 'Designer'
              assigned_hours:
                type: number
                default: 0
        technologies:
          type: array
          description: 'Array of technologies ids'
          items:
            type: string
        time_schedule:
          type: object
          properties:
            start_date:
              type: string
              description: 'It takes the current date as default'
            due_date:
              type: string
              description: 'It takes the current date as default'
            end_date:
              type: string
              description: 'It takes the current date as default'
            estimated_hours:
              type: number
              default: 0
            spent_hours:
              type: number
              default: 0
...
    Avatar:
      type: object
      required:
        - 'img_link'
      properties:
        img_link:
          type: string
          example: 'https://i.imgur.com/GCvrcOH.png'
paths:
...
  /projects:
    get:
      summary: Get all projects (Admin functionality)
      description: The projects retrieved autopopulate with the technology schema to get the technologies names. They also autopopulate with the user schema to give more information about the project members.
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Projects successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        403:
          description: This request requires an Admin role
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /projects/byUser:
    get:
      summary: Get all projects by the logged in user ID
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Projects successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /projects/project:
    get:
      summary: Get project by project ID
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: Project ID
          required: true
          type: string
      responses:
        200:
          description: Project successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      summary: Create new project
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        200:
          description: Project successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Update project by project ID
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: Project ID
          required: true
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        200:
          description: Project successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        403:
          description: The user must be a leader of the project or an Admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: Project not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete project by ID
      tags:
        - Project
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: Project ID
          required: true
          type: string
      responses:
        201:
          description: Project successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /avatars:
    get:
      summary: Get all avatars
      tags:
        - Avatar
      responses:
        200:
          description: Avatars successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      summary: Create new avatar
      tags:
        - Avatar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Avatar'
      responses:
        200:
          description: Avatar successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /avatars/randomAvatar:
    get:
      summary: Get random avatar
      tags:
        - Avatar
      responses:
        200:
          description: Avatar successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string